+++ bin/read_data.py
@@ -46,5 +46,13 @@ if __name__ == "__main__":
         dataset_id=args.sampleID,
     )
 
+    # Remove sampleID metadata from table:
+    if args.sampleID in spatialdata.tables['table'].uns.keys():
+        del spatialdata.tables['table'].uns[args.sampleID]
+
+    # Rename table into sample id
+    spatialdata.tables[f'{args.sampleID}_table'] = spatialdata.tables['table']
+    del spatialdata.tables['table']
+
     # Write raw spatialdata to file
     spatialdata.write(args.output_sdata, overwrite=True)